Move stats recorder out
